Begin section one

+++ spec/requests/api/v1/items_request_spec.rb
@@ -24,7 +24,7 @@ RSpec.describe 'Items API' do
       expect(item[:description]).to be_a(String)
 
       expect(item).to have_key(:unit_price)
-      expect(item[:unit_price]).to be_an(Integer)
+      expect(item[:unit_price]).to be_an(Float)
 
       expect(item).to have_key(:merchant_id)
       expect(item[:merchant_id]).to eq(id)
@@ -52,7 +52,7 @@ RSpec.describe 'Items API' do
     expect(item[:description]).to be_a(String)
 
     expect(item).to have_key(:unit_price)
-    expect(item[:unit_price]).to be_an(Integer)
+    expect(item[:unit_price]).to be_an(Float)
 
     expect(item).to have_key(:merchant_id)
     expect(item[:merchant_id]).to eq(id)
