+++ spec/requests/api/v1/merchants_request_spec.rb
@@ -89,7 +89,7 @@ RSpec.describe "Merchants API" do
       expect(item[:description]).to be_a(String)
 
       expect(item).to have_key(:unit_price)
-      expect(item[:unit_price]).to be_an(Integer)
+      expect(item[:unit_price]).to be_an(Float)
 
       expect(item).to have_key(:merchant_id)
       expect(item[:merchant_id]).to eq(id)
